feat: Update user shell in Authentik during shell edit
This commit adds functionality to update the user's shell in Authentik when editing the shell in the Nest Bot application. It retrieves the user's `pk` field from the database and makes a PUT request to the Authentik API to update the user's shell attribute. If the update fails, an error message is logged.