feat: Add /etc/shells file and test for shell filtering
This commit adds the /etc/shells file to the repository, which contains a list of valid login shells. It also includes a test file that checks if the shells retrieved from the get_shells utility function match the expected list of shells. This addition improves the reliability and accuracy of the shell filtering process in the Nest Bot application.