feat: add basic shell update functionality (#50)
* feat: add basic shell update functionality

* feat: add "Edit" button to shell block

* feat: get available shells from /etc/shells and use a drop down

* feat: Add pk field to users model

This is needed to update the user shell in authentik later

* feat: Update user shell in Authentik during shell edit

This commit adds functionality to update the user's shell in Authentik when editing the shell in the Nest Bot application. It retrieves the user's `pk` field from the database and makes a PUT request to the Authentik API to update the user's shell attribute. If the update fails, an error message is logged.

* chore: add a trim to each line

* feat: remove duplicate /usr/bin shells

* bug: fix shell_input to correctly replace /usr/ with /

* feat: Refactor shell_input to use async/await and get_shells utility function

This commit refactors the shell_input function in the shell_input.ts file to use async/await syntax and the newly added get_shells utility function. It replaces the previous implementation that used fs.readFileSync and multiple array operations to filter and transform the shells list. The new implementation improves readability and maintainability by separating the logic into a separate utility function. This change also ensures that the shells list is retrieved asynchronously, preventing potential blocking of the event loop.

* feat: Add /etc/shells file and test for shell filtering

This commit adds the /etc/shells file to the repository, which contains a list of valid login shells. It also includes a test file that checks if the shells retrieved from the get_shells utility function match the expected list of shells. This addition improves the reliability and accuracy of the shell filtering process in the Nest Bot application.

* feat: Update shells path in test.ts

This commit updates the path for retrieving shells in the test.ts file. The previous path was incorrect and has been corrected to "src/tests/shells". This change ensures that the shells are retrieved from the correct location, improving the accuracy of the test.